Extracted `mdbook build` to separate file/module
